requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
fastapi-cache2[redis]==0.2.1
polars==0.18.8
pyarrow==12.0.1
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
SQLAlchemy==2.0.19